+++ src/app/support-request/page.tsx
@@ -1,0 +1,9 @@
+export default function SupportRequestPage() {
+  return (
+    <div className="p-4">
+      <h1>Support Request Form</h1>
+      <p>Stepper component will be implemented here.</p>
+    </div>
+  );
+}
+